main/medical_wbc/src/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
numpy>=1.21.0